# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Noxware API
  version: 1.0.0
  description: |
    Accounts, subscriptions (NOWPayments crypto checkout), license keys, and the
    customer-support ticketing API (REST + inbound-email ingestion).

    ### Support intake over email
    MX/Email-Routing providers POST to `POST /api/mail/inbound` with header
    `X-Inbound-Secret: $MAIL_INBOUND_SECRET`. New mail opens a ticket; replies
    (Re:/Fwd:) append to the customer's most recent non-closed ticket.
servers:
  - url: https://api.noxware.cc
    description: Production
  - url: https://api-staging.noxware.cc
    description: Staging
  - url: http://127.0.0.1:3001
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/admin/users:
    get:
      tags: [Admin]
      summary: List/search users (admin only)
      parameters:
        - { name: search, in: query, schema: { type: string }, description: Matches email or username (LIKE %term%) }
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Users with role, ban state, subscription, ticket count
          content:
            application/json:
              schema:
                type: object
                required: [users, total]
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/AdminUser' }
                  total: { type: integer }
        '403':
          description: Not admin
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/users/{id}:
    patch:
      tags: [Admin]
      summary: Change role and/or ban/unban (admin only)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role: { type: string, enum: [customer, support, admin] }
                ban: { type: boolean }
                banReason: { type: string }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '403': { description: Cannot demote self / admins cannot ban admins }
        '404':
          description: Not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/users/{id}/grant:
    post:
      tags: [Admin]
      summary: Grant subscription days or lifetime access (admin only)
      description: |
        Extends from the later of now / current expiry, so grants stack.
        `lifetime:true` writes plan_id `lifetime` (~100 years) and is visible
        as "Lifetime" on the user's dashboard.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                days: { type: integer, minimum: 1, maximum: 3650 }
                lifetime: { type: boolean, default: false }
      responses:
        '200':
          description: Grant applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  subscription:
                    type: object
                    properties:
                      planId: { type: string }
                      expiresAt: { type: string, format: date-time }
                      lifetime: { type: boolean }
        '400':
          description: Neither valid days nor lifetime
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
  /api/admin/license-keys:
    post:
      tags: [Admin]
      summary: Generate unredeemed license keys (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                count: { type: integer, minimum: 1, maximum: 50, default: 1 }
                planId: { type: string, enum: ['1m', '3m', '6m', 'lifetime'], default: '3m' }
      responses:
        '201':
          description: Generated codes (prefix NOX1/NOX3/NOX6/NOXL)
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items: { type: string, example: 'NOXL-A1B2-C3D4-E5F6' }
  /api/admin/invites:
    get:
      tags: [Admin]
      summary: List invite codes with usage and expiry state (admin only)
      responses:
        '200':
          description: Invite codes, newest first
          content:
            application/json:
              schema:
                type: object
                required: [invites]
                properties:
                  invites:
                    type: array
                    items: { $ref: '#/components/schemas/Invite' }
    post:
      tags: [Admin]
      summary: Create an invite code (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string, minLength: 4 }
                maxUses: { type: integer, default: 1 }
                expiresInDays: { type: integer, nullable: true, description: 'Omit for no expiry' }
                note: { type: string, nullable: true }
      responses:
        '201':
          description: Created invite
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite: { $ref: '#/components/schemas/Invite' }
        '409': { description: Code already exists }
  /api/admin/invites/{id}:
    delete:
      tags: [Admin]
      summary: Revoke an invite code (sets expiry to now; keeps signup attribution)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200': { description: Revoked }
        '404': { description: Not found }
  /api/admin/users/{id}/revoke-invites:
    post:
      tags: [Admin]
      summary: Revoke all invite codes tied to a user (their signup code + codes they created)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Codes revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  revoked:
                    type: array
                    items: { type: string }
...
  /api/admin/users/{id}/reset-devices:
    post:
      tags: [Admin]
      summary: Staff reset of a user's devices (no cooldown, audited)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        '200':
          description: Devices cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  cleared: { type: integer }
                  nextAllowedAt: { type: string, format: date-time, nullable: true }
        '404': { description: User not found }
  /api/admin/audit-log:
    get:
      tags: [Admin]
      summary: Audit trail of admin actions (admin only)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 50, maximum: 200 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Newest-first audit entries
          content:
            application/json:
              schema:
                type: object
                required: [entries, total]
                properties:
                  entries:
                    type: array
                    items: { $ref: '#/components/schemas/AuditEntry' }
                  total: { type: integer }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT from /api/auth/login or /api/auth/register (header `Authorization: Bearer <token>`), valid 14 days.'
    inboundSecret:
      type: apiKey
      in: header
      name: X-Inbound-Secret
      description: Shared secret equal to the server's MAIL_INBOUND_SECRET.
  schemas:
...
    Invite:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        maxUses: { type: integer }
        uses: { type: integer }
        signups: { type: integer }
        note: { type: string, nullable: true }
        expiresAt: { type: string, nullable: true }
        expired: { type: boolean }
        exhausted: { type: boolean }
        createdAt: { type: string }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: 'Ticket is closed — open a new ticket' }
    User:
      type: object
      required: [id, email, username, role]
      properties:
        id: { type: integer, example: 7 }
        email: { type: string, format: email, example: jane@example.com }
        username: { type: string, example: jane }
        role:
          type: string
          enum: [customer, support, admin]
          description: Staff roles (support/admin) unlock the Staff endpoints.
        createdAt: { type: string, format: date-time }
...
    AdminUser:
      type: object
      required: [id, email, username, role, banned]
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        username: { type: string }
        role: { type: string, enum: [customer, support, admin] }
        banned: { type: boolean }
        banReason: { type: string, nullable: true }
        bannedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        subscription:
          nullable: true
          type: object
          properties:
            planId: { type: string }
            expiresAt: { type: string, format: date-time }
            active: { type: boolean }
        ticketCount: { type: integer }
    AuditEntry:
      type: object
      required: [id, action, createdAt]
      properties:
        id: { type: integer }
        actor:
          nullable: true
          type: object
          properties:
            id: { type: integer }
            username: { type: string }
            email: { type: string, format: email }
        action: { type: string, example: user.ban }
        targetType: { type: string, nullable: true, example: user }
        targetId: { type: integer, nullable: true }
        details: { type: object, nullable: true, additionalProperties: true }
        createdAt: { type: string, format: date-time }